refactor(docs): extract fuzzy search algorithm and improve TOC visibility
